Add new diagnostics.utils module
